+++ scripts/events/abilityCheck.js
@@ -10,7 +10,7 @@ function collectMacros(entity) {
     if (!macroList.length) return [];
     return macroList.map(i => custom.getMacro(i, genericUtils.getRules(entity))).filter(j => j);
 }
-function collectActorCheckMacros(actor, pass, checkId, options, roll, config, dialog, message) {
+function collectActorCheckMacros(actor, pass, checkId, options, roll, config, dialog, message, sourceActor) {
     let triggers = [];
     let effects = actorUtils.getEffects(actor);
     let token = actorUtils.getFirstToken(actor);
@@ -34,7 +34,8 @@ function collectActorCheckMacros(actor, pass, checkId, options, roll, config, di
             roll,
             config,
             dialog,
-            message
+            message,
+            sourceActor
         });
     });
     actor.items.forEach(item => {
@@ -56,7 +57,8 @@ function collectActorCheckMacros(actor, pass, checkId, options, roll, config, di
             roll,
             config,
             dialog,
-            message
+            message,
+            sourceActor
         });
     });
     if (token) {
@@ -81,14 +83,15 @@ function collectActorCheckMacros(actor, pass, checkId, options, roll, config, di
                 roll,
                 config,
                 dialog,
-                message
+                message,
+                sourceActor
             });
         });
     }
     return triggers;
 }
-function getSortedTriggers(actor, pass, checkId, options, roll, config, dialog, message) {
-    let allTriggers = collectActorCheckMacros(actor, pass, checkId, options, roll, config, dialog, message);
+function getSortedTriggers(actor, pass, checkId, options, roll, config, dialog, message, sourceActor) {
+    let allTriggers = collectActorCheckMacros(actor, pass, checkId, options, roll, config, dialog, message, sourceActor);
     let names = new Set(allTriggers.map(i => i.name));
     allTriggers = Object.fromEntries(names.map(i => [i, allTriggers.filter(j => j.name === i)]));
     let maxMap = {};
@@ -128,7 +131,8 @@ function getSortedTriggers(actor, pass, checkId, options, roll, config, dialog, 
                 roll: trigger.roll,
                 config: trigger.config,
                 dialog: trigger.dialog,
-                message: trigger.message
+                message: trigger.message,
+                sourceActor: trigger.sourceActor
             });
         });
     });
@@ -229,6 +233,27 @@ async function rollCheck(wrapped, config, dialog = {}, message = {}) {
     if (!returnData) return;
     let oldOptions = returnData.options;
     returnData = await executeBonusMacroPass(this, 'bonus', checkId, options, returnData, config, dialog, message);
+    let token = actorUtils.getFirstToken(this);
+    if (token) {
+        let sceneTriggers = [];
+        token.document.parent.tokens.filter(i => i.uuid !== token.document.uuid && i.actor).forEach(j => {
+            sceneTriggers.push(...getSortedTriggers(j.actor, 'sceneBonus', checkId, options, returnData, config, dialog, message, this));
+        });
+        let sortedSceneTriggers = [];
+        let names = new Set();
+        sceneTriggers.forEach(i => {
+            if (names.has(i.name)) return;
+            sortedSceneTriggers.push(i);
+            names.add(i.name);
+        });
+        sortedSceneTriggers = sortedSceneTriggers.sort((a, b) => a.priority - b.priority);
+        genericUtils.log('dev', 'Executing Check Macro Pass: sceneBonus');
+        for (let trigger of sortedSceneTriggers) {
+            trigger.roll = returnData;
+            let bonusRoll = await executeMacro(trigger);
+            if (bonusRoll) returnData = CONFIG.Dice.D20Roll.fromRoll(bonusRoll);
+        }
+    }
     if (returnData.options) genericUtils.mergeObject(returnData.options, oldOptions);
     if (message.create !== false) {
         genericUtils.mergeObject(messageData, {flags: options.flags ?? {} });